Further cleanup and refactoring

+++ Fox/src/Fox/Core/Types.hpp
@@ -32,7 +32,7 @@ namespace fox
 	using byte_t     = fox::uint8_t;
 	using size_t     = fox::uint64_t; //Number of bytes
 	using offset_t   = fox::uint64_t; //Number of bytes offset
-	using count_t    = fox::uint64_t; //Number of elements
+	using count_t    = fox::uint32_t; //Number of elements
 
 	using uuid_t     = std::bitset<128>;
